requirements.txt
asgiref==3.5.0
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.12
chart-studio==1.1.0
click==8.1.0
colorama==0.4.4
dash==2.3.1
dash-auth==1.4.1
dash-bootstrap-components==1.0.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.3
django-heroku==0.3.1
filelock==3.6.0
Flask==2.1.0
Flask-Compress==1.11
Flask-SeaSurf==0.3.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy #==1.22.3
pandas #==1.4.1
pipenv==2022.3.28
platformdirs==2.5.1
plotly==5.6.0
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
retrying==1.3.3
six==1.16.0
sqlparse==0.4.2
tenacity==8.0.1
tzdata==2022.1
ua-parser==0.10.0
urllib3==1.26.9
#virtualenv==20.14.0
virtualenv-clone==0.5.7
Werkzeug==2.1.0
whitenoise==6.0.0
zipp==3.7.0
accelerate==0.26.1
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.7.0
anyio==4.6.0
asn1crypto==1.5.1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
configparser==7.1.0
contourpy==1.1.1
cryptography==43.0.1
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
distro==1.9.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==2.1.0
filelock==3.16.1
fonttools==4.43.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kaleido==0.2.1
kiwisolver==1.4.5
langchain==0.3.1
langchain-core==0.3.6
langchain-text-splitters==0.3.0
langsmith==0.1.129
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.9.1
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.0
openai==1.51.2
openpyxl==3.1.5
orjson==3.10.7
packaging==23.2
pandas==2.2.3
parso==0.8.4
pdfminer.six==20240706
Pillow==10.0.1
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.0.0
pytz==2024.2
pywin32==308
pyxlsb==1.0.10
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
safetensors==0.4.5
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowflake-connector-python==3.12.2
snowflake-snowpark-python
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.35
stack-data==0.6.3
streamlit==1.39.0
sympy==1.13.3
tenacity==8.2.3
toml==0.10.2
tomlkit==0.13.2
toolz==1.0.0
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.3
watchdog==5.0.3
wcwidth==0.2.13
XlsxWriter==3.2.0
yarl==1.13.0

